refactor: separate controller and router
